power: qpnp-smb5: Fix -Wvoid-pointer-to-int-cast
"drivers/power/supply/qcom/qpnp-smb5.c:309:16: warning: cast to smaller
integer type 'u8' (aka 'unsigned char') from 'const void *'
[-Wvoid-pointer-to-int-cast]" error, forbidden warning: qpnp-smb5.c:309

+++ drivers/power/supply/qcom/qpnp-smb5.c
@@ -306,7 +306,7 @@ static int smb5_chg_config_init(struct smb5 *chip)
 {
 	struct smb_charger *chg = &chip->chg;
 	struct device_node *node = chg->dev->of_node;
-	int subtype = (u8)of_device_get_match_data(chg->dev);
+	int subtype = (uintptr_t)of_device_get_match_data(chg->dev);
 
 	switch (subtype) {
 	case PM8150B:
